feat: Add columns for type and target time in routine_exercise table
This commit adds two new columns, "type" and "target_time", to the routine_exercise table in the public schema. The "type" column is of type character varying(20) and has a default value of 'steps-reps'. The "target_time" column is of type time without time zone. The commit also adds constraints to validate the values of these columns.

The changes were made in the following files:
- scripts/migrations/time_based_exercises.sql
- src/commons/infra/db/kysely-pg/db.ts
- src/routine/domain/models/routine-exercise.ts
- src/routine/infra/repository/kysely-routine-exercise-repository.ts